+++ plan/dag_plan_test.go
@@ -16,12 +16,8 @@ package plan_test
 import (
 	. "github.com/pingcap/check"
 	"github.com/pingcap/tidb"
-	"github.com/pingcap/tidb/context"
-	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/parser"
 	"github.com/pingcap/tidb/plan"
-	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/testleak"
 )
 
@@ -663,54 +659,6 @@ func (s *testPlanSuite) TestDAGPlanBuilderAgg(c *C) {
 		c.Assert(err, IsNil)
 		p, err := plan.Optimize(se, stmt, is)
 		c.Assert(err, IsNil)
-		c.Assert(plan.ToString(p), Equals, tt.best, Commentf("for %s", tt.sql))
-	}
-}
-
-func (s *testPlanSuite) TestDAGPlanEmptyTable(c *C) {
-	defer func() {
-		testleak.AfterTest(c)()
-	}()
-	store, err := newStoreWithBootstrap()
-	c.Assert(err, IsNil)
-	testKit := testkit.NewTestKit(c, store)
-	defer func() {
-		store.Close()
-	}()
-	testKit.MustExec("use test")
-	testKit.MustExec("drop table if exists t, t1")
-	testKit.MustExec("create table t (c1 int)")
-	testKit.MustExec("create table t1 (c1 int)")
-	testKit.MustExec("analyze table t, t1")
-	tests := []struct {
-		sql  string
-		best string
-	}{
-		{
-			sql:  "select * from t where t.c1 <= 50",
-			best: "TableReader(Table(t)->Sel([le(test.t.c1, 50)]))",
-		},
-		{
-			sql:  "select * from t where c1 in (select c1 from t1)",
-			best: "SemiJoin{TableReader(Table(t))->TableReader(Table(t1))}(test.t.c1,test.t1.c1)",
-		},
-		{
-			sql:  "select * from t, t1 where t.c1 = t1.c1",
-			best: "LeftHashJoin{TableReader(Table(t))->TableReader(Table(t1))}(test.t.c1,test.t1.c1)",
-		},
-	}
-	for _, tt := range tests {
-		ctx := testKit.Se.(context.Context)
-		stmts, err := tidb.Parse(ctx, tt.sql)
-		c.Assert(err, IsNil)
-		c.Assert(stmts, HasLen, 1)
-		stmt := stmts[0]
-		is := sessionctx.GetDomain(ctx).InfoSchema()
-		err = plan.ResolveName(stmt, is, ctx)
-		c.Assert(err, IsNil)
-		err = expression.InferType(ctx.GetSessionVars().StmtCtx, stmt)
-		c.Assert(err, IsNil)
-		p, err := plan.Optimize(ctx, stmt, is)
 		c.Assert(plan.ToString(p), Equals, tt.best, Commentf("for %s", tt.sql))
 	}
 }
